Added temp sword and fixed a few errors

+++ Summer Wave Game/Assets/Scripts/Main Character/PlayerDamageMath.cs
@@ -26,11 +26,11 @@ public class PlayerDamageMath : MonoBehaviour {
 		newDmg = 0;
 	}
 
-	public void setBaseDmg(float dmg){
+	public void setBaseDmg(int dmg){
 		baseDmg = dmg;
 	}
 
-	public float getNewDmg(int level, int mod, int cL10M, int l50M, int l99M){
+	public int getNewDmg(int level, int mod, int cL10M, int l50M, int l99M){
 		newDmg += baseDmg + level * mod + (level / 10) * cL10M;
 
 		if(level >= 50){
